+++ src/main/java/roomescape/domain/ThemeRepository.java
@@ -2,7 +2,7 @@ package roomescape.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import roomescape.service.exception.ResourceNotFoundCustomException;
+import roomescape.domain.exception.ResourceNotFoundCustomException;
 
 @Repository
 public interface ThemeRepository extends JpaRepository<Theme, Long> {
